fix the compiler warnings
None of these change behaviour, all of them predate this branch:

 - rtp.cpp: the write() on the rtcp socket was meant to wake the receive
   thread out of poll() when stop() is called, but the socket is only
   bound, never connected, so the write always failed with EDESTADDRREQ
   and no datagram was ever sent. The thread leaves through the 1000 ms
   poll timeout instead, which is why this never showed. Dropped.
 - vtuner.cpp: in the committed diseqc switch, voltage was printed in the
   debug line although neither of the two branches may have assigned it.
   Initialised, and the line now says '?' for that case instead of
   claiming horizontal. The uncommitted switch declared a second voltage
   that is not used at all.
 - session.cpp: unused rtp_port.
 - config.cpp: m_settings was initialised before members that are
   declared ahead of it.
 - config.cpp, rtsp.cpp: signed against unsigned comparisons made
   explicit; NO_STREAM_ID_FILTER is unsigned and m_rx_data_pos is
   compared against a size_t.

+++ rtsp.cpp
@@ -237,7 +237,7 @@ again:
 			return RTSP_OK;
 		}
 		size_t len2 = 4 + (((unsigned char)m_rx_data[2] << 8) | (unsigned char)m_rx_data[3]);
-		if (m_rx_data_pos < len2)
+		if ((size_t)m_rx_data_pos < len2)
 		{
 			return RTSP_OK;
 		}
